pyproject.toml
[project]
name = "gerador-de-apostila"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.123.4",
    "google-cloud-storage>=2.0",
    "google-generativeai>=0.8.5",
    "langchain>=1.1.0",
    "langgraph>=1.0.4",
    "langgraph-checkpoint>=3.0.1",
    "psycopg2-binary>=2.9",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0",
    "uvicorn>=0.38.0",
]
